requirements.txt
db-sqlite3
langchain
langchain_openai
python-dotenv
fake_useragent
pandas
numpy
scikit-learn
python-dotenv
playwright
langchain_community
flask
flask_cors
fetchai
logging
uagents
uagents-core
flask[async]
graphrag
neo4j
hypercorn
quart
quart_cors